feat: Security and CORS

+++ planner/routes/events.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter,HTTPException,status,Body,Depends,Request
-from sqlmodel import select
+from planner.auth.authenticate import  authenticate
 from beanie import PydanticObjectId
 
 from planner.database.connection import Database
@@ -28,20 +28,28 @@ async def retrieve_event(id: PydanticObjectId) -> Event:
     return event
 
 @event_router.post("/new")
-async def create_event(body: Event) -> dict:
+async def create_event(body: Event,user: str = Depends(authenticate)) -> dict:
+    body.creator=user
     await event_database.save(body)
     return {
         "message": "Event created successfully"
     }
 
 @event_router.delete("/delete/{id}")
-async def delete_evnet(id: PydanticObjectId) -> dict:
+async def delete_event(id: PydanticObjectId, user: str = Depends(authenticate)) -> dict:
+    event_user = await event_database.get(id)
+    if event_user != user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Oeration not allowed"
+        )
     event = await event_database.delete(id)
     if not event:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Event with supplied ID does not exist"
         )
+
     return {
         "message": "Event deleted successfully"
     }
@@ -59,7 +67,13 @@ async def delete_all_events() -> dict:
     }
 
 @event_router.put("/{id}",response_model=Event)
-async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
+async def update_event(id: PydanticObjectId, body: EventUpdate,user: str = Depends(authenticate)) -> Event:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation not allowed"
+        )
     update_event = await event_database.update(id,body)
     if not update_event:
         raise HTTPException(
